2.19: converted most of the upgrade info to 64-bit

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/quest/Pickaxe.java
@@ -259,7 +259,7 @@ public class Pickaxe extends MeleeWeapon {
 	}
 
 	public String upgradeAbilityStat(long level){
-		int dmgBoost = 8 + 2*level;
+		long dmgBoost = 8 + 2*level;
 		return augment.damageFactor(min(level)+dmgBoost) + "-" + augment.damageFactor(max(level)+dmgBoost);
 	}
