+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/rings/RingOfSharpshooting.java
@@ -52,7 +52,7 @@ public class RingOfSharpshooting extends Ring {
 	@Override
 	public String upgradeStat1(long level) {
 		if (cursed && cursedKnown) level = Math.min(-1, level-3);
-		return Integer.toString(level+1);
+		return Long.toString(level+1);
 	}
 
 	@Override
